Factor out the code for checking if an image has already been uploaded
